moed laout handler logic into it's own class.
